requirements.txt
flask==3.1.0
anthropic==0.52.0
requests==2.32.3
beautifulsoup4==4.13.3
python-dotenv==1.1.0
